+++ server_op/socket/bind/bindsocket.h
@@ -6,7 +6,7 @@ namespace http {
   class bindsocket: dnnsocket {
   private:
   public: 
-    bindsocket(int domain, int service, int protocol, int port, u_long bridge)
+    bindsocket(int domain, int service, int protocol, uint16_t port, uint32_t bridge)
     : dnnsocket(domain, service, protocol, port, bridge){};
 
     int connect_to_network(int sock, struct sockaddr_in address);
