+++ server_op/socket/connect/connectocket.cpp
@@ -1,6 +1,6 @@
 #include "./connectsocket.h"
 
-http::connectsocket::connectsocket(int domain, int service, int protocol, int port, u_long bridge) 
+http::connectsocket::connectsocket(int domain, int service, int protocol, uint16_t port, uint32_t bridge) 
 : dnnsocket(domain, service, protocol, port, bridge) {
   set_connection(connect_to_network(get_socket(), get_address()));
   test_connection(get_connection());
