+++ server_op/socket/connect/connectsocket.h
@@ -6,7 +6,7 @@ namespace http {
   class connectsocket: dnnsocket {
   private:
   public: 
-    connectsocket(int domain, int service, int protocol, int port, u_long bridge);
+    connectsocket(int domain, int service, int protocol, uint16_t port, uint32_t bridge);
     int connect_to_network(int sock, struct sockaddr_in address);
   };
 }
